Removed color of JoystickView parent Container
Removed color parameter for Container widget containing child = JoystickView

Reason: Explicitly specifying white color to the parent Container widget is not required. This results in color of the Container being transparent which is expected by default. So when changing the Scaffold color to any other color say black, the square Container area around the JoystickView should inherit the color of the parent widget. 

This just makes the example code cleaner.